+++ proiect/src/main/java/com/example/demo/domain/UserLimit.java
@@ -2,6 +2,7 @@ package com.example.demo.domain;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -11,9 +12,9 @@ public class UserLimit {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int userLimitId;
-    private int maxAmountPerTransactionRon;
-    private int maxDailyAmountRon;
-    private int maxDailyTransactionsCount;
+    private BigDecimal maxAmountPerTransactionRon;
+    private BigDecimal maxDailyAmountRon;
+    private BigDecimal maxDailyTransactionsCount;
     private String status;
     @Temporal(TemporalType.TIMESTAMP)
     private Date createdAt;
